+++ Surface_mesh_approximation/examples/Surface_mesh_approximation/vsa_approximation_example.cpp
@@ -3,7 +3,7 @@
 
 #include <CGAL/Exact_predicates_inexact_constructions_kernel.h>
 #include <CGAL/Polyhedron_3.h>
-#include <CGAL/VSA/approximate_mesh.h>
+#include <CGAL/Surface_mesh_approximation/approximate_mesh.h>
 
 #include <CGAL/Polygon_mesh_processing/orient_polygon_soup.h>
 #include <CGAL/Polygon_mesh_processing/polygon_soup_to_polygon_mesh.h>
