+++ Surface_mesh_approximation/test/Surface_mesh_approximation/vsa_class_interface_test.cpp
@@ -7,7 +7,7 @@
 
 #include <CGAL/boost/graph/graph_traits_Polyhedron_3.h>
 #include <CGAL/Variational_shape_approximation.h>
-#include <CGAL/VSA/L2_metric_plane_proxy.h>
+#include <CGAL/Surface_mesh_approximation/L2_metric_plane_proxy.h>
 
 typedef CGAL::Exact_predicates_inexact_constructions_kernel Kernel;
 typedef Kernel::FT FT;
